requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-debug-toolbar==4.0.0
redis